fixed new problem with tests... moved test
