+++ unity-client/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/Tests/UUIDComponentTests.cs
@@ -1114,60 +1114,6 @@ namespace Tests
         }
 
         [UnityTest]
-        public IEnumerator OnPointerHoverFeedbackIsBlockedByUI()
-        {
-            DecentralandEntity entity;
-            BoxShape shape;
-            InstantiateEntityWithShape(out entity, out shape);
-            TestHelpers.SetEntityTransform(scene, entity, new Vector3(8, 2, 10), Quaternion.identity, new Vector3(3, 3, 3));
-            yield return shape.routine;
-
-            var onPointerDownModel = new OnPointerDown.Model()
-            {
-                type = OnPointerDown.NAME,
-                uuid = "pointerevent-1"
-            };
-            var component = TestHelpers.EntityComponentCreate<OnPointerDown, OnPointerDown.Model>(scene, entity,
-                onPointerDownModel, CLASS_ID_COMPONENT.UUID_CALLBACK);
-            Assert.IsTrue(component != null);
-
-            yield return null;
-
-            DCLCharacterController.i.SetPosition(new Vector3(8, 1, 7f));
-
-            var cameraController = GameObject.FindObjectOfType<CameraController>();
-
-            // Rotate camera towards the interactive object
-            cameraController.SetRotation(45, 0, 0);
-
-            yield return null;
-
-            var hoverCanvasController = Environment.i.interactionHoverCanvasController;
-            Assert.IsNotNull(hoverCanvasController);
-
-            // Check hover feedback is enabled
-            Assert.IsTrue(hoverCanvasController.canvas.enabled);
-
-            // Put UI in the middle
-            UIScreenSpace screenSpaceShape =
-                TestHelpers.SharedComponentCreate<UIScreenSpace, UIScreenSpace.Model>(scene,
-                    CLASS_ID.UI_SCREEN_SPACE_SHAPE);
-            yield return screenSpaceShape.routine;
-
-            UIContainerRect uiContainerRectShape =
-                TestHelpers.SharedComponentCreate<UIContainerRect, UIContainerRect.Model>(scene,
-                    CLASS_ID.UI_CONTAINER_RECT);
-            yield return uiContainerRectShape.routine;
-
-            yield return null;
-
-            // Check hover feedback is no longer enabled
-            Assert.IsFalse(hoverCanvasController.canvas.enabled);
-
-            DCLCharacterController.i.ResumeGravity();
-        }
-
-        [UnityTest]
         public IEnumerator OnPointerHoverDistanceIsAppliedCorrectly()
         {
             DecentralandEntity entity;
